requirements.txt
dnspython==1.12.0
requests==2.7.0
